refactor(theme): extract ThemeContext and types, improve theme toggling logic
