+++ src/theme/useTheme.tsx
@@ -1,19 +1,9 @@
-import React, { useState, useEffect, createContext, useContext } from "react";
-
-export type Theme = "dark" | "light" | "system";
-
-interface ThemeContextType {
-  theme: Theme;
-  resolvedTheme: "dark" | "light";
-  setTheme: (theme: Theme) => void;
-  toggleTheme: () => void;
-}
-
-const ThemeContext = createContext<ThemeContextType | undefined>(undefined);
+import React, { useState, useEffect, useContext } from "react";
+import { ThemeContext } from "./ThemeContext";
+import type { Theme } from "./types";
 
 export const ThemeProvider: React.FC<{ children: React.ReactNode }> = ({ children }) => {
   const [theme, setTheme] = useState<Theme>(() => {
-    // Initialize theme from localStorage or default to system
     if (typeof window !== 'undefined') {
       const savedTheme = localStorage.getItem('theme') as Theme | null;
       return savedTheme || 'system';
@@ -39,7 +29,6 @@ export const ThemeProvider: React.FC<{ children: React.ReactNode }> = ({ childre
     }
   }, [theme]);
 
-  // Persist theme preference
   useEffect(() => {
     if (typeof window !== 'undefined') {
       localStorage.setItem('theme', theme);
@@ -51,7 +40,7 @@ export const ThemeProvider: React.FC<{ children: React.ReactNode }> = ({ childre
   }, [resolvedTheme]);
 
   const toggleTheme = () => {
-    setTheme((t) => (t === "dark" ? "light" : t === "light" ? "system" : "dark"));
+    setTheme((t: Theme) => (t === "dark" ? "light" : t === "light" ? "system" : "dark"));
   };
 
   return (
@@ -61,6 +50,7 @@ export const ThemeProvider: React.FC<{ children: React.ReactNode }> = ({ childre
   );
 };
 
+// eslint-disable-next-line react-refresh/only-export-components
 export function useTheme() {
   const ctx = useContext(ThemeContext);
   if (!ctx) throw new Error("useTheme must be used within a ThemeProvider");
